[roll] Roll fuchsia [audio][drivers] Add VIM3 support
We configure the aml-g12-tdm driver for buses A and B in the VIM3.
Bus A goes into the BT chip and bus B goes into the GPIO header where
we can probe the signals.
The aml-g12-tdm driver is slightly modified to support the A311D SoC,
which is very similar to the A905D2 already supported. The driver's
GPIO usage is made optional to support the VIM3 case.

Original-Fixed: 132037

Test: aml-g12-tdm-test
Test: device-enumeration-test on VIM3
Test: Measurement of signals on the VIM3 GPIO headers with oscilloscope.
Test: Loopback of signals on the VIM3 GPIO header by shorting output and
input. Signal generated via audio-driver-ctl as in:
$ audio-driver-ctl -d 17f0e62d tone
and captured in another fx shell with:
$ audio-driver-ctl -d 9a74ca44 record /tmp/t

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888515
Original-Revision: 26a9efc3d81c76aed21e9c92c378d9d3fb14a72d
GitOrigin-RevId: 3e481da85e60415a64358bf418eddc02d5b13c36
Change-Id: I9112368a6141d03ad2c94d3a6f52f0982c279947