[roll] Roll fuchsia [zxdb][starnix] Disable e2e test on LTO builds.
This test fails if the execute_syscall function gets inlined which
is prevented with an #[inline(never)] annotation in most builds.
It seems that the annotation is not sufficient to prevent inlining with
LTO when the size of the function crosses LLVM's heuristic bounds, so
for now we'll disable this test on LTO builders.

Original-Bug: b/293716995
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901362
Original-Revision: 40a05fb862d48274ec8b4ba6c5000770d6980482
GitOrigin-RevId: 5eba256532c0664ef21d0c53daea6f190b4a25a6
Change-Id: I5fe40552fb0def85d4f6657571e5c6b0008bfb43