[roll] Roll fuchsia [storage][inspect] Replace volume in lazy node callbacks.
Currently when a volume is unmounted the lazy node stays unless or until
the related volume is deleted. The data in it is just blank. After
remount despite re-registering the volume, this state continues.

Instead of re-inserting the LazyNode which doesn't replace the callback
on the existing lazy node, nor replace it for the caller, we just
replace the volume in a map that is looked up at call time.

Original-Bug: 130564
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889954
Original-Revision: 96216aa0e6d1afa85e46606751eb6941ac17d053
GitOrigin-RevId: 6783929d2b2147bf348e207c48f337339e44705b
Change-Id: I836e25a8fbb431d51c643d8cb6d0a69853d49465