[roll] Roll fuchsia [elfldltl] Introduce loadinfo-mapped-memory API
`LoadInfoMappedMemory` implements the Memory API, encapsulating a memory
object and the `LoadInfo` associated with it.

`LoadInfoMappedMemory.ReadArray` will find the loaded segment that
corresponds to the passed in `ptr`. The offset in the memory object is
calculated using the segment's information, and is forwarded to the
`ReadArray` call on the underlying memory object.

Original-Bug: 118060
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886698
Original-Revision: 80049ec06f2666c788bc5c027cf4a8d1c661b7e9
GitOrigin-RevId: c236422b4c09788cc1c946ec68f244d40469d480
Change-Id: Id72b12a3c56696d94ae8de2c23cb8b8dacb0187f