[roll] Roll fuchsia [starnix] Use UserAddress as key for private futexes
When we added support for shared futexes, we changes private futexes to
use the VMO koid + offset as a key. That change was incorrect because if
you have the same file mapped twice into your address space, the private
futexes on those two mappings are separate. This CL makes FutexTable
have two modes, one for shared futexes and one for private futexes. The
shared mode uses koid + offset and the private mode uses UserAddress as
keys.

It's difficult (if not technically impossible) to create a robust test
for this change because we would need to test that the futexes in one
mapping are not woken up when we direct a wake towards the futexes in
another mapping. However, a futex can always have a spurious wakeup, so
the test cannot actually fail if we wake up both sets of futexes.

However, a non-robust test case (see
https://fuchsia-review.googlesource.com/c/fuchsia/+/901558) shows that
this new behavior is correct.

Original-Fixed: b/295928846
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901658
Original-Revision: 4783e14d8651fdfc0e08e482fd4b3ee56a63666b
GitOrigin-RevId: f799958b747897e8a7303d4d50f08a7c0fd63115
Change-Id: I050eff24f4e1f771c75cd1a50193d5a8d3cba2a0