[roll] Roll fuchsia [storage][benchmarks] Flush journal for Fxfs cold benchmarks.
We clear cache by remounting the filesystems, but in Fxfs case
remounting includes replaying the journal back into the in-memory layer,
so the items appear to be recently modified and "hot" again anyways.
This overrides the cache clearing process to include forcing a flush of
the journal by filling it up until inspect reflects that a flush has
occurred.

Original-Bug: 130564

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888250
Original-Revision: 7109db965d680ade20092bc92898c1bad87f90ed
GitOrigin-RevId: 6d05aa2b9cbcfa1d0f4d05a97d6a2d87c44ecd5b
Change-Id: Ia6a31e78913dad0defd53e8ee6eb31b210363086