[roll] Roll fuchsia [ld] Add Zircon version of LdStartupSpawnProcessTests
This uses fdio_spawn to launch the test process in the normal way
via the system program loader (i.e. the fuchsia.process.Launcher
FIDL srevice).

This is not ready to be used yet, so for now it's just compiled
and not tested.  Before it can be used, it will need plumbing for
a fuchsia.ldsvc.Loader service to deliver the dynamic linker VMO,
as well as for the startup dynamic linker to correctly link the
vDSO references in the test modules.  Manual local testing has
verified that it works as far as component_manager logging that
it couldn't launch the PT_INTERP because there was no ldsvc
handle supplied.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900579
Original-Revision: 0d159b3ed9c48ecaa17dcb9e7ac5bfdbef3a9358
GitOrigin-RevId: 416e392bcd67a8af1616ab6bffc6416ea8dab19b
Change-Id: I7021b5da9da275c5db8c15a6bd0648ffb1b43691