[roll] Roll fuchsia [sys] Start of a bedrock program type
The purpose of a bedrock program type is to describe which operations
are supported on a program, that corresponds to the invariants we want
to maintain.

With the exception of "start" that will be addressed later, each
operation is safe to use by both component manager or bedrock users in
general. There's no component manager only backdoor or play favorites.

Right now Program is a minimal wrapper over ComponentController but we
should be able to move other program-y API here over time, such as an
API to get the outgoing directory of the program.

Original-Bug: 122024
Multiply: fuchsia-pkg://fuchsia.com/component-manager-tests#meta/component_manager_tests.cm
Multiply: component-manager-stress-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897620
Original-Revision: 1ec41f1e6b286c5314d8d7fdb43ff2059975d9e0
GitOrigin-RevId: 4260d671fc2217e78007875d2ab19878d6fba7a8
Change-Id: I0e8fb28ef5c96ecff3a3f32f8405c3e6f8b54c8c