[roll] Roll fuchsia [ufs] Fix a SCSI transfer length bug
This CL fixes an issue where the 'Transfer Length' in SCSI and the
'Expected Data Transfer Lengt' in UPIU do not match.
The problem was setting the request's task_tag and ExpectedDataTransferLength
after memcpy the request to the descriptor buffer.

Original-Bug: 124835
Test: ufs-unit-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895438
Original-Revision: 844d6ea87cb7e678a4fab4056593c4739950210d
GitOrigin-RevId: 52225670b431fbd174cd400890da4a4174adca9d
Change-Id: I77fcfa10b913376a2aee3fe8da8e4284c3d96ffb