[roll] Roll fuchsia [components] Create logger without going through namespace
StartAction creates a logger attributed to the started component.
This requires a connection to the LogSink protocol routed to the
component.

Previously, StartAction got a connection in a round-about way by
constructing the component's namespace, finding the appropriate entry
with a directory that contains LogSink, and opening the protocol
through that directory.

Now, StartAction issues a route request for the LogSink protocol use,
without connecting through the namespace.

This simplifies StartAction and removes the need to write a "connect
to protocol" method in Namespace, at least for now. ELF runner still
connects to LogSink in the component's namespace in the same way.

Original-Bug: 131717
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897618
Original-Revision: 9135ed9c7a916f17e544c47c27bb3a647804e239
GitOrigin-RevId: 9cbd24911f0f7737ee74dd0c73bd7c4b45a6a7f0
Change-Id: I0ab30f25568ce6b11ac0464e8968f7c7657c1361