[roll] Roll fuchsia [icu][assembly] Fixes up the intl assembly
It turns out that some products have an existing
configuration that is not expressible in the current
assembly config for the `intl` subsystem.

This change handles that issue by introducing
the missing configuration option. The actual
solution would probably be to reduce the config
variation, but that's a more extensive change
that I'd like to keep separate.

Original-Bug: 117974, b/296136021
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902176
Original-Revision: 78c0b4b74a59c341f3f3f8ceaafab50dc3cb1272
GitOrigin-RevId: 555173b6ba7f1976c6b4277ad09e965aecee746c
Change-Id: I01a66a957f86da3e726206fa1ae8333d8efe4d91