[roll] Roll fuchsia [sdk] Move driver testing out of core_testing
These libraries are needed in the bazel sdk, which does not include the
core_testing library. Remove the testonly gn arguments and add the sdk
library in to the full core sdk to be included in the bazel sdk.

Original-Bug: 127797
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902400
Original-Revision: 405feb8589e928d6bfa3a9ec03c3125c0776d73a
GitOrigin-RevId: f9cf5c5b637083d31080a76c8ddac89fd51f841f
Change-Id: Iec853edee8ffa85cc8604bc55f4d694683cf19bc