[roll] Roll fuchsia [fdf][examles] Remove demo driver base test
This test is way too simple to add any value. The driver was not doing
anything except a log message on start. The test was not validating
anything. The demo number driver that exists here has a lot more value
and there is nothing here that the demo number one doesn't have. So we
can remove this example/test.

Original-Bug: 132131
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903400
Original-Revision: 09eac7a88092da980b8945972000ab63ad3c4fd9
GitOrigin-RevId: 278c022c6836f3532ba7db48799bc3c01ab29c8b
Change-Id: I3411092b191a4ce5d5555b5251b206b54d197bcf