[roll] Roll fuchsia [diagnostics] Drain the socket when not measuring
The benchmark is sufficiently CPU intensive that other threads can
impact the benchmarks. Prior to this CL, a thread is spawned to drain
the socket to ensure it doesn't overflow, but this thread could affect
timings. (See https://fxrev.dev/901921 for an example of a change which
one would expect to not have an impact on the benchmark, but it actually
causes a regression.)

The change here makes sure draining of the socket happens outside of the
measurement.

Original-Fixed: 131474
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902618
Original-Revision: 0a75ea8f1d55df021c5c1ef47942a50c9d1fd154
GitOrigin-RevId: b29d6219cd169b195cc99d832b87eed9f042ecad
Change-Id: I355a1fc15d51700c7bbea68218d22efb18e59cdd