[roll] Roll fuchsia [profiler][fidl] Add Targeting Components And Tests
This adds support to the profiler FIDL protocol for specifying component
and test targets.

Profiling a component will cause each task it launches to be profiled.
Profiling a test will launch a test via test manager and attach to it
via its job.

The actual implementation to do so is stubbed out for now and will
follow in coming patches.

This adds ComponentConfig and TestConfig as a type of TargetConfig. It
also pushes the vector into specifically launching adhoc Tasks as
attempting to profile multiple components or tests at once is somewhat
complicated and there isn't an immediate use case for it.

Original-Bug: b/293459772
Original-Bug: b/294081668
Original-Bug: b/284190829
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872191
Original-Revision: 57c773779757a20c2b57393f0e039a8b2ea18450
GitOrigin-RevId: 73e8e641f8984cc92b7a1a6ecd92d96a0dad0ae5
Change-Id: Ic48716fdfa53cc109864215ce21ab34df56633b5