[roll] Roll fuchsia [ld] Fix ld-unittests in instrumented builds
This fixes a couple of problems in instrumented builds.

The build rules for the test data needed to account for the
libprefix an instrumented variant toolchain might use in the
PT_INTERP.  This is already handled implicitly for the Zircon
case, but the explicit host_test_data() plumbing for the Linux
case needed to change.  Also, the compile-time plumbing for the
prefix to use for in-process tests that already existed for
Zircon needed to be used for the Linux version as well.

At runtime, the Linux version had the same issue seen on Zircon
with write-protecting the bss when functions already in progress
still touch the counters there via calculating with an old cached
__llvm_profile_counter_bias value taken at function entry.  So
the posix-startup.cc code needed the same workaround that Zircon
uses to disable the write-protection when using llvm-profdata.

A previous change had disabled the spawn-process tests altogether
since more work is needed for the Zircon version to work.  This
also re-enables them for the Linux version so these cases are
again tested to prevent any regressions creeping in before the
Zircon spawn-process tests are ready to be enabled.

Original-Fixed: 132169
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903339
Original-Revision: a3060f44b581216a5de2ec8d708aadaac4f4c2c4
GitOrigin-RevId: c9593eee31708423887b0e45b4347c110903d4cf
Change-Id: I9d98f6f343a10a0e72b8a0453992e43c39ba4358