[roll] Roll fuchsia [bedrock] Support standalone namespace entries in serve_processargs
This serves the needs of "/pkg" directories and directory/storage
capabilities and the like. You can add an Open capability (or something
that is TryIntoOpen) at a path, and the Open capability will be opened
when someone opens that path.

A subtle requirement here is that we need to delay exercising the open
capability until the user writes a message onto the namespace entry, as
seen in `add_directory_helper` in component manager. That has been
replicated here and in fact doesn't affect regular namespaced objects.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900279
Original-Revision: 36064bb7056f8746af15a264898ca88b17e4f959
GitOrigin-RevId: 3d780bb5276ae8c79da479e35e665d208c53d153
Change-Id: Ifa422001c175a26dc9abde3b3505676775e67eb7