[roll] Roll fuchsia [heapdump][ffx] Print a usage error without the "internal error" preamble
This error is not a bug, so the "BUG: An internal command error
occurred" preamble is confusing to the user. By using ffx_bail, such a
preamble is not printed.

In addition, this makes this error consistent with the other user-facing
errors generated by this source file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903259
Original-Revision: 2c2ae511988bb1a9fbf6ceabf69e63d4a908b31c
GitOrigin-RevId: 86a220fc7e482357f04c5ebec4e18139810fdb5c
Change-Id: I86a0c9bbef8d776d4a63c61c7c92db712d463426