[roll] Roll fuchsia [icu_data] Remove unneeded target
This target was only used at one place, and was
a forwarding target anyways. So we can remove it.

I noticed this late in working on
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/903498,
and I didn't want to destroy the presubmit results to
add this inconsequential change.

Original-Bug: 117974
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904098
Original-Revision: ae4f085a6c8f1b0793e07b093447c959bdc585d6
GitOrigin-RevId: fd1b3068b3ac308aecec95b5c7505afd64f90dfd
Change-Id: I6160aedcf94ac1b9dc369a59cf24e29ec7013d3f