[roll] Roll fuchsia [netstack3] Use a single ID for ICMP sockets.
Note that this change does not guard against bugs if you use a wrong
type of socket ID. The rationale is that this is a transitory state
and the refactor to use `DatagramSocketSpec` to implement ICMP sockets
will change most of the existing functions. There is currently also no
usage of ICMP sockets elsewhere, so keeping the throwaway work minimal
as long as the tests are passing.

Original-Bug: 125482
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902099
Original-Revision: ebdff69d77b94abff8b70da4dd179ad10a8d4c52
GitOrigin-RevId: 4bfa8839be9a1d835d0ad2c39a2b2b69666249a7
Change-Id: I51ca8a0e5becbe35e51785f68464c4ad1750f75f