[roll] Roll fuchsia [netstack3] Add UNREACHABLE neighbor state
Add the UNREACHABLE state for dynamic neighbor entries, which is
conceptually similar to the PROBE state: packets continued to be sent to
the cached link address, and neighbor probes are actively sent to
resolve the neighbor. However, in UNREACHABLE, the probes are multicast
rather than unicast, to allow for discovering a link-layer address
change for the neighbor.

Probes are transmitted as long as packets continue to be sent to the
neighbor, but with exponential backoff rather than on a fixed retransmit
timer. When packets are not sent, the netstack also stops sending
probes.

Original-Bug: 35185
Test: netstack3-core-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898678
Original-Revision: 829351e91fb5987887c2d890aec6f986fbb956b9
GitOrigin-RevId: 51e2e8cc582feb0da563b559877dc1d66c3413a9
Change-Id: I3434478a0a7b9b87012a8172db04ce3f09819d62