[roll] Roll fuchsia [rust-3p] Update to 5.2.0
This updates the notify crate to 5.2.0. Unfortunately while it looks
like they attempted to include the license file, it appears that it
doesn't actually include the license text, so we'll need to maintain our
own fork for now.

Along the way, we also need to update our windows-sys to 0.45.0. However
we still have a few things still using windows-sys 0.42.0, so it keeps
an empty version of that crate as well until we finish migrating off
of it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904140
Original-Revision: d327fbe5357cb86be936dc0fdcfa4f76862fed7b
GitOrigin-RevId: 069610f2088cfa0147597afe71e7f723f4ea1a00
Change-Id: I20f6818679bd1bd8ac1b4a3752f8b88d4eb60852