[roll] Roll llvm-project/libc [libc] Do not run tests on vendor implemented math
We currently remap vendor implementations of math functions to provide a
temporarily functional `libm.a` for the GPU. However, we should not run
tests on any files that depend on these vendor implementations as they
are not under our control and are not always present.

The goal in the future is to remove the need for this by replacing all
the vendor functionality, but for now this is a workaround.

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D158213

GitOrigin-RevId: 50e54c7e248f5412de61056584d244d534341690
Original-Revision: debfcfe845d51bb608d3a7a5b4b29068a93aa55f
Change-Id: I3766a3b0f1f8623c0dffbc8b846bee73fe6d416e