[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Do not run tests on vendor implemented math
We currently remap vendor implementations of math functions to provide a
temporarily functional `libm.a` for the GPU. However, we should not run
tests on any files that depend on these vendor implementations as they
are not under our control and are not always present.

The goal in the future is to remove the need for this by replacing all
the vendor functionality, but for now this is a workaround.

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D158213

GitOrigin-RevId: 683f55010f1955bf53b09760a88238aa9d0a5a23
Original-Revision: debfcfe845d51bb608d3a7a5b4b29068a93aa55f
Roller-URL: https://ci.chromium.org/b/8772466602117436433
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904359
Original-Revision: b02a2ade2fac6c37dc9b27efc9f681c9785b51ad
Change-Id: I893dc5894c5ad7980d38fe28a0dc63ac5d5517b7