[roll] Roll fuchsia [ffx] Add debugging for ssh connection failure
To understand fxb/127900 better, we need some debugging on the
initial ssh connection.  Depending on what information comes out
of this, we can improve the ssh handling (including a more
refined approach to timing out on the parsing the ssh output).

This also adds an ssh timeout so we get a useful report if we hang
indefinitely.

Original-Bug: 288918487
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903500
Original-Revision: a2afdb1ef03b5673b9a652eb55f859b6f60579c6
GitOrigin-RevId: 60a58163b1bfc45d249de9c8073cedeb02ef2b11
Change-Id: Ie092566371d9dcbb7ef381a76d170b18a4d1bf42