[roll] Roll fuchsia [netstack3] Ensure SyncCtx outlives NonSyncCtx
If NonSyncCtx outlives SyncCtx, we can encounter a panic due to
referency IDs (while SyncCtx is being torn down, it expects to
hold the sole copy of referency IDs). This is a particularly thorny
footgun in tests.

To avoid this, we avoid impling Clone on NonSyncCtx, and instead expose
accessor methods for the SyncCtx and NonSyncCtx on the bindings Ctx
struct, which itself is Clone.

Test: netstack3-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903940
Original-Revision: 05b846c46a23123591e66c14a6f94cf673bccab7
GitOrigin-RevId: a12317ad5035d5b9d2d449eae4ca71c6804d6a8f
Change-Id: Icd760b9fbccfe3977a55a2ffdfd45179c01c5e37