[roll] Roll fuchsia [assembly] add a FileRelativePathBuf crate
This crate allows the deserialization of paths that are relative
to the file they're given in, and then resolving those paths
against the path to that containing file.

It also allows the conversion of paths into a file-relative form,
given that the path to the file and the path to the containing
file are both given from the same base.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903503
Original-Revision: 7ac10c388b384937e59134758ae25a9c45de75f7
GitOrigin-RevId: 48941d5c89b39cc9a22799d352f61fed3ea0114e
Change-Id: I709a6fa42918f173997d91c2a3d73e53e2c54865