[roll] Roll fuchsia [elfldltl] Add SegmentWithVmo and AlignedRemoteVmarLoader
This enhances the elfldltl::LoadInfo API template to take an
optional wrapper template that's be used to enhance the various
LoadInfo::*Segment types with additional functionality.

elfldltl::VmarLoader now uses a new template class parameterized
by the LoadInfo type, elfldltl::VmarLoader::SegmentVmo<LoadInfo>.
This can be specialized to override how each segment is mapped,
replacing the file VMO with another and choosing the VMO offset
and whether the VMO is mutable or should be copied to map a
writable segment.

The new elfldltl::SegmentWithVmo API in the new header file
<lib/elfldltl/segment-with-vmo.h> expliots these new extension
points.  It allows new instantiations of LoadInfo where each
segment (except ZeroFillSegment) can hold its own VMO handle.

The new elfldltl::AlignedRemoteVmarLoader variant of VmarLoader
requires that each DataWithZeroFillSegment be normalized so that
the boundary between data and zero-fill is page-aligned; it will
not handle partial-page zeroing.  To use this loader, a LoadInfo
can be prepared with elfldltl::SegmentWithVmo::AlignSegments.  It
installs a per-segment VMO with the partial-page tail of the data
portion zeroed in the VMO contents.  This can be used with an
elfldltl::SegmentWithVmo::NoCopy instantiation of LoadInfo for
one-off dynamic linking, or with elfldltl::SegmentWithVmo::Copy
either for the full zygote model or simply to reuse the prepared
VMO before relocation.  Only the NoCopy case is tested so far.

Original-Bug: 91206
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900838
Original-Revision: 003b2e620d9860e2bd8b8f57731a35fb8b7f77fd
GitOrigin-RevId: fdfda41d5eb78a944a91eaa0b4a848698ddc051e
Change-Id: Ia42d3db2de9797bd3b4a36688d65bac0ea32e7be