[roll] Roll fuchsia [zircon][qemu] fix serial port when using the -g option to fx qemu
A previous refactor had broken the automatic redirection of serial
to the console when using the -g switch, which enables a graphics
window.

Generally refactor the code dealing with the -serial switch to be a bit
more robust. Notably --no-serial overrides everything, and -S defaults
to 'stdio'.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904400
Original-Revision: 11818dad91f79e8f9452d033d5e53624e9017332
GitOrigin-RevId: da386be6a6b713550e8b7feae501b398112e9a93
Change-Id: Id969de620ea55ada85d8d698316857a5906ec6d1