[roll] Roll fuchsia [vm][test] Poll VMO committed bytes
The DropParentCommittedBytes test verifies that the child gets
attributed for visible pages once the parent is dropped. When the test
is run as a component though, dropping the parent handle might not
necessarily destroy the parent. It is possible for memory_monitor to be
querying the parent attribution count at the same time holding it alive.

So poll the committed count a few times, sleeping in between, until it
eventually matches the expected value. In practice, this should not take
long as the attribution query should be pretty quick. If something goes
wrong though, the test timeout will kick in and terminate the test.

Added a helper to do the polling so that other tests that have a similar
pattern can reuse it.

Original-Bug: 132120
Multiply: fuchsia-pkg://fuchsia.com/core-vmo-test-package#meta/core-vmo-test-package.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903124
Original-Revision: 931620cf36c0d8ebcf478935f20a6c8f5441734b
GitOrigin-RevId: e3bab6103f89f20eb834e2c19269d2ab8f69ee29
Change-Id: I3362b296a92db5bf4ae5bdbe59d0aed0b2fb49fd