[roll] Roll third_party/pigweed pw_build: Fix output_logs_ unused warning
This variable is only used when being built with the right toolchain.
In order to set this variable in fuchsia and not have GN complain it's
unused, it must be set as not_needed in other situations.

Original-Bug: b/295524695
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/166991
Original-Revision: 10f09568e72d08db8a301ad20d03cea70a833f0b
GitOrigin-RevId: 17d1cd8284d737b8df213378dbf098a6b5f6c3bb
Change-Id: I7857aac13220593b87a5de0b9d2c589f465ddd28