[roll] Roll fuchsia [log][rust] Provide an option for blocking log writes
Instead of dropping a log when the socket buffer is full, this change
provides an option to wait and retry when the socket becomes writable
again.

It's believed that most clients will prefer performance at the cost of
dropping some logs when too many are being generated too fast. However,
for those clients that wish to guarantee their logs make it to the
Archivist at the potential cost of some performance an option is offered
to retry indefinitely when the socket buffer is full. Although no
performance impact has been observed in the current target use case.

Multiply: diagnostics-log-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904178
Original-Revision: 5d208997f3ad41015e81f672d4033382a55497d1
GitOrigin-RevId: d18dd00e2f975ec3009e38b8edc8eeee54b360a8
Change-Id: Iaeaa2dd2984cf71b44a60d6e7a374768b4475c1f