[roll] Roll fuchsia [cm_rust] runner is not optional
If a program declaration is present, it must specify the runner.
This property is enforced by cm_fidl_validator.
So there's no value in making it optional in cm_rust.

This way we'll be able to say that a component has a runner if and only
if the program block is present.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905398
Original-Revision: 048252648793b6cffbe1bbb0849c7c6a30a0d6fc
GitOrigin-RevId: 9f1559b8afed1576b8a68903365a8f9ce13b727a
Change-Id: Ic833b53ec843e8c8dd399072a6dc041f5fb3ec51