[roll] Roll fuchsia [assembly] Remove vestiges of legacy intl configuration
Required firming up ICU config as well, to emphasize that
there are products that are not interested in the intl
config. Mostly this is the "basic" products.

Furthermore, added the defaults for each of intl, setui
and icu to be used if no special values are needed.

Original-Bug: 117974
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905711
Original-Revision: 84dd188493475f44cbf054af24810da0709da1ae
GitOrigin-RevId: ac96ab4bb932ec0d15bde00ef60421b0d905ad26
Change-Id: Ie8ca1b0538cc59210408ecc95b1a443219bd58b0