[roll] Roll fuchsia [ufs] Support large block size I/O
This CL allows passing a data buffer of up to 1MiB in a single
write/read command. To achieve this, the size of the `UTP Command
Descriptor` has been increased to 8KiB and the number of PRDT entries
has been increased to 256.

Original-Bug: 124835
Test: ufs-unit-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900159
Original-Revision: 85f9f0b6409b6370b6ecf1966a69126994e389fb
GitOrigin-RevId: 7c9ca4009246f3859fcf0a80693bef224a944f49
Change-Id: Ia1f55f1a3310b2c337e4c6588820cbbc397cfc9e