[roll] Roll fuchsia [netstack3] Add dual stack conn state
Encode whether a connected socket is connected in `ThisStack` or the
`OtherStack` using an enum. This provides additional type safety when
performing dual stack operations because it allows us to match on the
`DualStackConnState` and explicitly handle the dual-stack case.

As a fallout from this CL, we now can clearly see the remaining TODO's
associated with dual-stack datagram sockets.

Original-Fixed: 132333

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903098
Original-Revision: 542a17be56fdc44e48e51c4328cf79a03164f702
GitOrigin-RevId: dfbe4e3b211d68c84cb47ed0e0a90772cd8fe625
Change-Id: Id76e116acd1f659dc5c3b6e3448a784d644c3901