[roll] Roll fuchsia [components] Don't set default path in testing decl builders
Capability decls builders used in tests were setting arbitrary
default source_paths that are not the documented default
of "/svc/{name}" for protocols and services.

Setting a default value in the decl makes the false assumption that
the test will add it that to the outgoing dir, which is not true
for tests that don't use RoutingTest.

For protocols, the path "/svc/foo" is added by default to the outgoing
directory only in RoutingTest, defined in CM. However, the decl builders
are defined in the cm_rust library and can be used in any other test.
A generic library should not depend on details of a specific test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904364
Original-Revision: e67c2390d141fee90d8c8ada173d104052179c17
GitOrigin-RevId: c16ac3d7b14964efeb22e98668d331058d0f458f
Change-Id: I7f16a3f3403cae9ec7d599ba59d4bab344a8a528