[roll] Roll fuchsia [zircon][tests] Move thread-initial-state-test into core-tests
This is a low-level test of basic kernel semantics and should be
included in the core-tests suite.  This gives it bringup-only and
boot-test versions of the test to catch kernel issues earlier in
simpler circumstances, in addition to the component version of
the test that is easier to use and debug in userland.

Original-Bug: 132495
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906483
Original-Revision: 90c2a1b8ab14d6ad8929a2b9f1b89641080c2cda
GitOrigin-RevId: d6754a13dd0cd499d92a705219f34135f0134807
Change-Id: Ib9faecf9e4cc8f49ace2439510a3fce2187f778e