[roll] Roll fuchsia [assembly] Ramdisk mode is enabled via CLI arg
Instead of placing the ramdisk mode in the product config, we now pass
it to product assembly as a CLI argument. This will allow us to use the
same product config between products even if they opt to put the
packages in a ramdisk in the ZBI (such as netboot).

Original-Bug: b/297090553
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903905
Original-Revision: 0767e813ae67dc49c8ec028324ada0a11e861406
GitOrigin-RevId: 3b287645fc735c899dad2b3f7bb9db8a52a5183d
Change-Id: I98a1f8853e6f58176beb2e680a316aab06d89da0