[roll] Roll llvm-project/libc [libc] Remove `MAX_LANE_SIZE` definition from the RPC server
This `MAX_LANE_SIZE` was a hack from the days when we used a single
instance of the server and had some GPU state handle it. Now that we
have everything templated this really shouldn't be used. This patch
removes its use and replaces it with template arguments.

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D158633

GitOrigin-RevId: e9d6d519accbf6c73608519e41ea090604d24496
Original-Revision: e6b91af5a746f79d95dfb07249d869533c43d25a
Change-Id: Iec4cb44c150692675bf3027d1806440b706aa1f9