[roll] Roll fuchsia [net-types] Transpose nested witness types
Adds a `transpose` method on `Witness` that allows nested witness types
to be transposed. For example, given
`addr: NonMappedAddr<SpecifiedAddr<T>>`, `addr.transpose()` gets you a
`SpecifiedAddr<NonMappedAddr<T>>`.

As an implementation detail, a `TransposableWitness` trait is introduced
that is implemented for all single-layer `Witness` impls in order to aid
type inference. However, because the `transpose` method is on `Witness`
and not `TransposableWitness`, callsites will not have to bring the
`TransposableWitness` trait into scope.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906369
Original-Revision: 3b0551112a970712bfe22a379de436931c24ddc7
GitOrigin-RevId: 757792744c9f53a63978f9ee0f6ee50694b5d908
Change-Id: I634216ad46a70a19e2988873cc8f20745e8b886b