[roll] Roll fuchsia [elfldltl] elfldltl::SegmentWithVmo::GetMutableMemory
This is an object to wire up elfldltl::LoadInfoMutableMemory using
elfldltl::SegmentWithVmo so that a new copy-on-write VMO child is
created for each mutable segment as needed.  This returns a Memory
object using a modified elfldltl::MappedVmoFile to map in each new
VMO so relocations can be applied to its contents.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900480
Original-Revision: 066b946dfd6e3661f271de1eeb307ab3126a5983
GitOrigin-RevId: e6874a0419705df88235a2ca91f681c264f529ba
Change-Id: Ie52dca5abb69e8067a02a28e106a3aa4692cc98b