[roll] Roll fuchsia [trace][rust] Option to block provider init until actually initialized.
Because the trace bindings spawn a separate thread to run trace provider
logic it's possible for the trace events early in a Rust program's
execution to be missed even if a trace is already running when it starts.

This change offers an API that blocks the provider creation bindings
from returning until the spawned thread has had a chance to process any
trace start/stop messages that were received while synchronously
registering the provider.

We can't make this the unconditional behavior because some Rust programs
are dependencies of trace_manager and deadlock when trying to
register themselves, it's mostly useful for integration tests anyways.

Original-Bug: b/293716995

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906422
Original-Revision: 8e992b4daa13409341cdf39787e058da48452342
GitOrigin-RevId: 94b08ab3b5c864856bd97155c85b187a03fc219e
Change-Id: Ic80621b94c4afaa353cb21626e7d6977371da2bc