[roll] Roll fuchsia [logs] Delete log_ng
This is now redundant and can be removed. The code in this plugin
was copied to the previous "ffx log" plugin due to FIDL APIs not being in the SDK and eventually will be converted away from a dual-mode plugin
when it is possible to do so and all FIDL dependencies have been
exported to the SDK and versioned.

Original-Bug: 121413
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907323
Original-Revision: f26512657205c4eea4c7b449c9141a515e86b56c
GitOrigin-RevId: 035c1c862bee66dcfe6d7da310ab1d0defae3791
Change-Id: I3808e42fcf24b4daa3268e1337b6febe035fee71