[roll] Roll fuchsia [zbi][rust] Ensure that zither's rust ZBI bindings are built
This change begins to generate the crate of zither's rust bindings for
the ZBI data types - and does so as a first step toward replacing the
hand-maintained //src/sys/lib/fuchsia-zbi/abi with them.

Fixed in this change is a backend bug that generated rust 'modules' with
improper names (i.e., those that feature an hyphen). The ZBI library
was the first occurrence of that bug, as it attempted to map
driver-config.fidl to driver-config.rs. The fix is just to map
hyphens to underscores.

Original-Bug: 132599
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906433
Original-Revision: feef1e25e65634d8f94c9eb47b903f6031d1b517
GitOrigin-RevId: 1d289f0dcf7b7f43768811e5c8f019e41d856a43
Change-Id: I2ea97ef1858ffad9cae0512429dd26cca9ddcfb1