[roll] Roll fuchsia [fxtest][rewrite] Implement --parallel
This implements the --parallel flag in a way that is not compatible with
the original fx test. Originally it would control the per-suite parallel
flag, but now it controls the number of simultaneous suites run on a
single target.

The ffx test --parallel flag overrides the suite's settings, but it is
most commonly used to ensure that suites that have internal crosstalk do
not fail to run. We should always respect the tests.json parallel flag,
which is what this CL does for the new fx test.

Original-Fixed: b/295340779
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904204
Original-Revision: 5e602087681f6aa81f7e7478fd2b326248353793
GitOrigin-RevId: b343a33d7b0511aa34b7f88a0b62b53dad660c00
Change-Id: I45c02d15bc0c115689804a87768d501e6a8d1838