[roll] Roll fuchsia [bt] Stop storing unparsed advertising data.
Parse advertising data and EIR data as soon as we receive it.
Add flags to the parsed advertising data.
Refactor DiscoveryFilter to use the parsed data, which will be null if
the data failed to parse.

Add services that are in the EIR for BR/EDR peers to the peer's
services.

Original-Bug: b/294861923
Original-Bug: fxbug.dev/85368
Test: fx test bt-host-gap-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895781
Original-Revision: e1be5032063de971fc6202210c5fa7a980ecc599
GitOrigin-RevId: be9ed36a6816d737870fdf5644a91b2636e9137e
Change-Id: I17d81dfecb431eb2641ac3e639fe99b56eda8b59