[roll] Roll fuchsia [ui][a11y] avoid spurious zoom when panning
When trying to pan the magnified view, it's possible
to intentiionally adjust zoom level as well.

This happens when the two fingers are used to pan,
since even a small change in the absolute distance
can cause a large change in scale. (The scale is
multiplied by |new_distance/old_distance|.)

This happens more often for me on small displays,
since there is less room to spread my fingers
apart to begin with.

To avoid spurious zoom adjustments, we check the
distance between the fingers in the previous
frame of the gesture. If the fingers are close
together, we ignore the scale adjustment
implied by the change in finger distance.

A user may still pan and zoom simultaneous in
general. And specifically, the user may switch
from pan-only (with fingers close together),
and pan+zoom without lifting their fingers off
of the screen. As soon as the distance between
fingers exceeds a threshold, the next pointer
sample with a change in finger distance will
cause an adjustment to the zoom factor.

Note that this change only affects persistent
magnification mode. In temporary magnification
mode, there's no way to change the zoom level
anyway.

While there: improve the debugging output of
the `ExpectThatTransformIs` helper, by including
the line number of the code that called the
helper.

Original-Bug: b/293890717
Test: unit tests, manual

Manual test: placed two fingers close together on a
7" touchscreen, and verified that I could pan the
magnified view without distracting changes to zoom.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905932
Original-Revision: 185a2cfc38ae22ddf25b0c358e1d0cdafef0fa30
GitOrigin-RevId: 6aafee9c27212c1fb50daa332368a05c86d9ff09
Change-Id: Iaa0c3ae57d839bd011324e2b3856179b22ad3ae8