[roll] Roll fuchsia [build] Relocate drivers-build-only and create_all_drivers_doc to //bundles/drivers/BUILD.gn
This moves these targets out of the root //bundles/BUILD.gn, so that
they are not included in the build graph unless added by a recipe,
developer, or tests.

This reduces the size of the GN build graph that is created for
most developers that are not adding all tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907734
Original-Revision: f591d2be96bd1d1167d46c7e19f70033fc5cd1c2
GitOrigin-RevId: ede786ed68c25f7ad533e19c403c9cdabe59e247
Change-Id: Ib1568cbd4a0bd5c9605cfa7051259ad3a0d5a103