[roll] Roll fuchsia [starnix] Change the filter_methods macro to a proc macro
The macro has it was written has become increasingly complex and is not
correct for template parameters with bounds.

Instead of trying to fix this by increasing the complexity once again,
change the macro into a proc macro that can use the syn crate to do the
parsing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907742
Original-Revision: 234bbd5bacdd334d8c354279d21dfdd2e42fe52d
GitOrigin-RevId: 26f06dcd40b0221425a06940aab8b2c9fe054c4f
Change-Id: I2125cf00436ed0d4d6b99ebf47af87babf29812d