[roll] Roll fuchsia [bazel-tool] Add 'actions' command.
This command can print the actions generated for a target, or
set of targets. This is similar to doing a plain `bazel aquery`
except that inputs will be filtered to remove the _very_ long
lists of sysroot and toolchain related input files, which pollute
the output and make it completely unreadable.

Original-Bug: None
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871157
Original-Revision: 04988ab050ffc69693d4112febc3d3c24ddb77f8
GitOrigin-RevId: 74f8d2d68f457f2321813a5bc8bc3ed2f11956dd
Change-Id: I07d30a65af78476b1b3f699584269345d1104b11