[roll] Roll infra/recipes [rust_toolchain] Fix bootstrapping with an upstream compiler
This complexity is unfortunate, but until we can teach upstream to
correctly build targets against our hermetic sysroot, it's probably the
best we can do.

This re-introduces the use of `-Ztarget-applies-to-host` for upstream
bootstrapping only. We aren't using that in production, but it's useful
for local builds, or if we miscompile our compiler (something that seems
to happen frequently when LTO is enabled).

Test: ( source fuchsia-env.sh && ./x.py --config fuchsia-config.toml build compiler )
Original-Bug: 108610
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/901566
Original-Revision: d5191a34cee3b48f9ad2d992422e420b0018c2a2
GitOrigin-RevId: 46549e68823d5d2a457b9edcbb9d84fe39634bed
Change-Id: Ib6251c637e5a59c6d42428bc1b3d16fb947f9b25