[roll] Roll fuchsia [session] Route `dev-bt-hci` to session
The `dev-bt-hci` directory capability is used to publish bt-hci
devices. This allows the Bluetooth stack to communicate with the
Bluetooth controller.

Route this directory capability to session so that it can be accessed
from Starnix.

Route this directory to the tiles-session which is the default element
for the session.

Original-Bug: b/278287238
Test: fx reboot and verify capability is accessible from starnix
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903123
Original-Revision: c852947fab38eb52d6994b33b16849e2d8886041
GitOrigin-RevId: 6b9d100b2791f4ba62bfb74f5ea1ff7dd45184c2
Change-Id: I7d91a57d63b7f0222fbfd4357cd00d816670ab7d