[roll] Roll fuchsia [kernel][hotplug] Move hotplug deadline into tests
arch_mp_cpu_hotplug currently waits up to 5 seconds for a core to come
online. This change modifies the function to not wait at all, and
instead rely on the caller to wait on the core. This required modifying
the test code to wait instead.

Original-Fixed: 130793
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903479
Original-Revision: 50182fc634243a2ed4eb6ef0a8931970c2add262
GitOrigin-RevId: fe766977662d4fe2508f054930cf5041d81311d5
Change-Id: I8caae4f70862cc00b82efaf114b448f7b1185e16