[roll] Roll llvm-project/libc [libc] Add more test cases to the argument list tests
This patch adds some extra cases to the existing argument list test in
`libc`, mainly dealing with arguments of varying sizes and primitive
types.

The purpose of this patch is to provide a wider test area when we begin
to provide varargs support on the GPU as there is no other runtime code
that really tests it. So, running these tests more exhaustively in the
GPU libc project will serve as the runtime tests for GPU vararg support in
D158246.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D158867

GitOrigin-RevId: fce65f7807e136c8ef2904cc8f5e0798bb5880a6
Original-Revision: 09e54466b7bdb016c4ba6c00d520b6e5064281d8
Change-Id: I6f33c2c1a32b4694a72cc4162c40c0e5814a0b88