[roll] Roll fuchsia [fxtest][rewrite] Overhaul --fuzzy matching
It turns out tha Jaro-Winkler similarity is not a great fit for the use
case of selecting tests. The main reason for this is its emphasis on
string prefixes, which for device tests are all
"fuchsia-pkg://fuchsia.com/". This results in direct selection attempts
to return far more than intended (b/297885060).

This change overhauls fuzzy matching and implements the --fuzzy flag to
control it (matching the CLI for the old implementation).

Instead of using Jaro-Winkler, we switch to Damerau–Levenshtein
distance, a specialization of Levenshtein distance in which character
transpositions are counted as an edit distance of 1 instead of 2. This
helps with common typos in test names while more closely matching the
behavior of the old fx test.

Original-Fixed: b/297885060
Original-Fixed: b/295340076

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908874
Original-Revision: b662e06a84ebcd9f91321654db3f038aa9f34ed4
GitOrigin-RevId: fd3f2274388db729ecc63577ba4ee918b404031a
Change-Id: I0b04d703cff703a99dbce5d74d43b86997ca8254