[roll] Roll fuchsia [netstack3] Confirm reachability on TCP handshake, ACK
Provide an upper-layer confirmation to NUD that the netstack has forward
reachability when either a TCP handshake completes or at least one byte
of data is acknowledged by a remote peer. In this case, the neighbor
that we are confirming is reachable is either the peer on the TCP socket
(if link-local) or the link-local gateway to that peer.

This change causes a small regression in some of the netstack3 TCP
socket benchmarks over network-tun, about 1-4% for some of the larger
message sizes, due mainly to the cost of performing a routing table
lookup in the TCP segment receive path. This could likely be optimized
by implementing route caching in the future.

Original-Bug: 35185
Test: netstack3-core-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907603
Original-Revision: e012064cda59298f73c7de87cae07cbce67c9985
GitOrigin-RevId: 30f921e81eaf654620b7ec62dc81ee0e702cdfe4
Change-Id: I856b7d4864c7a9ade1c3405848337e998a8b9db2