[roll] Roll fuchsia [zircon][clock] Add ZX_RIGHTS_SIGNAL to ZX_DEFAULT_CLOCK_RIGHTS
Allows a clock creator to create clocks with signaling allowed. It is the
clock creator's responsibility to attentuate this right appropriately.

This changes the outcome of attempting to signal on a clock. Before this
change, you would get ZX_ERR_ACCESS_DENIED if you try to signal
ZX_CLOCK_STARTED on a clock. Now, it is ZX_ERR_INVALID_ARGS, since now
signals are allowed in general but ZX_CLOCK_STARTED is not allowed in
particular.

Modified the documentation for `zx_clock_create()` to match.

Original-Bug: b/297433088
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908656
Original-Revision: b6b29f6cd0f1dff377e12288979ae8643c7f1111
GitOrigin-RevId: f993307f0d5984aea86ca2fb7f7a3b97e5d4987e
Change-Id: I8185b45b0fcde0785489530cd5eab875ebfe4902