[roll] Roll fuchsia [fuchsia-async] Efficiently spawn detached tasks
For tasks that are going to be detached, using Task::spawn wraps the
future in a Fuse and a RemoteHandle that won't be used.

This change adds spawn methods to EHandle specifically for spawning
detached tasks. All of the uses of Task::spawn_on were using detach so
it has been replaced with calling EHandle::spawn_detached.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906306
Original-Revision: 7c22e710e4183a6782b9a5574dbb765e4e764495
GitOrigin-RevId: f997e5b118ed8336f3f8a75083bad31c73abe2aa
Change-Id: I0369c47876a3344c4ed76cbdbadd5ac1560c031b