[roll] Roll fuchsia [threads][test] Text fixture to clean up leaked mappings
The Threads core tests map the thread stacks in the root VMAR, but do
not unmap when done. Use a custom text fixture instead to perform the
clean up. The text fixture creates a sub-VMAR inside the root VMAR to
use for any mappings, and on teardown this sub-VMAR is destroyed.

Original-Bug: 130559
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907468
Original-Revision: e2dfbb56a385db4b68c52f095ddddebe55639922
GitOrigin-RevId: 668a7f675d5fb7f3be7ea208914169da56b1036e
Change-Id: I6fd4e8c33b4bbf4e2d2441b22a2320e6ba040942