[roll] Roll llvm-project/libc [libc] Remove hard dependencies on linux from macros headers
Currently all of this logic expects to include and link with the headers
in the `linux/` directory. This patch adds a wrapper macro to optionally
add the appropriate dependency if it exists. This is preliminary to
adding some GPU specific versions of these files.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D159021

GitOrigin-RevId: c8a6b16c3373dcb1163c115d4f218660a01d365b
Original-Revision: 43b397d92f83ad56ab44d03c882fe3ea6428faa1
Change-Id: Iecd0413e81c71072357a2ce49352d3bd582f8a4d