[roll] Roll fuchsia [audio][drivers][simple-codec] Fix race when creating server instances
Previously simple-codec-server acquired the instances lock after
creating a new server instance and binding it to the dispatcher. If this
happened on a non-dispatcher thread, there was a race where the instance
could receive OnUnbound on the dispatcher before it was added to the
instances list, causing a crash (or assert failure on debug builds). To
fix this, make simple-codec-server acquire the instances lock before
creating a new instance, thus ensuring that it is always added to the
list before OnUnbound is able to run.

Original-Fixed: 132750
Multiply: nelson-brownout-protection-test: 1000
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909561
Original-Revision: fc0adc81e85241ddc964d06aa0cdbb4e5a361821
GitOrigin-RevId: 326b58eca675a8635d097b3dc7aae4346e804ea0
Change-Id: Ie0f5f47630328339921cf20b17e473a2461a9837