[roll] Roll fuchsia [mdns] Properly clean up ResponderPublisher on error from callback.
While investigating b/295351837, it was noticed that multiple queries
for a subtype of a service (for example) would eventually cause the
services that were not a part of that subtype to no longer appear to be
available on the network. This was because `GetPublicationNow` method
was terminating early if the `OnPublication` callback indicated
anything other than success.

This would prevent the `OnGetPublicationComplete()` from executing,
which was necessary to ensure that the accounting state was being
updated properly.

This change makes sure that the `OnGetPublicationComplete()` method
gets called even in the event of a non-success response from
`OnPublication`.

Original-Bug: b/295351837

Test: unit test added
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909500
Original-Revision: c05e078cea2483d6c653b8fc24945cd257c677c4
GitOrigin-RevId: 1f30f85d311672c66ab20346dca25fc1dccc7b7c
Change-Id: I464f20d77ad2a3fd2423e5e88466503e2619de92