[roll] Roll fuchsia [flatland] ObjectLinker cleanup.
For DX convenience, we previously gave ObjectLinker the ability to
dispatch link resolution/invalidation to a particular thread.  However,
we found that in practice, this resulted in race conditions that made it
difficult to deterministically test the code; we wanted only *part* of
the closure to run on a separate thread.  Consequently, we stopped using
ObjectLinker's built-in redispatch, and added our own in LinkSystem.

This CL removes ObjectLinker's redispatch ability: enough time has
passed that we are confident that it won't be useful in the future.

Test: gfx_unittests, flatland_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735283
Original-Revision: 36eb57b0a04e7d3d3e5457c5416437e02c1f7c2f
GitOrigin-RevId: 55ec6ecc834391e5c6d3db9aeb8ec57e592a26c8
Change-Id: I02d61b0ade1db13ba180fdae1e08d439cc3e1bc6