[roll] Roll fuchsia [fxtest][rewrite] Create new process groups for job handling
Without this change, sending a signal to the processes only affects the
initial processes created, but does not affect their children. This is
fine in many cases, but specifically when running an `fx` command it
results in killing only `fx`, and not the command being run by `fx`.

The result is that AsyncCommand ends up hanging until the process stops
emitting output, which is typically when it decides to terminate
anyway.

This change instead creates a new process group for all commands, and
when terminating or killing it signals to the entire process group. The
result is that both the calling script (`fx`) and the subprocess are
sent the same signal.

Original-Fixed: b/297439693

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908658
Original-Revision: 2467f8af6b02663b77b7a93c3386cec2b8562fa5
GitOrigin-RevId: e5a41aaa709024bc7bce4169ca0017c6e5fdfe14
Change-Id: I03a35d57cef4c4b42b6288cd54232d5d4d00422f