[roll] Roll fuchsia [gen-cargo] Improve 3p crate feature handling
Previously we recorded the list of features specified in
third_party/rust_crates/Cargo.toml for each 3p crate, and when a crate
was listed as a dep we'd add it's corresponding list of features to that
dep entry.

This fails to properly handle a workaround we use for tokio due to
fxbug.dev/106164. We specify it with an empty features list, and then
add `--cfg=feature=<foo>` flags to the individual GN targets for
different platforms so that we can have a separate set of features on
host and target. When a target depended on tokio, the generated
cargo manifest would incorrectly ask for tokio with no features and
often failed to compile.

To fix this I'm completely doing away with the process of reading
features from the original cargo manifest. Instead we'll solely gather
the list of features by parsing rustflags on the generated GN targets
looking for `--cfg=feature=`. This simplifies the logic a bit and fixes
the current rustdoc failures were seeing which were caused when
fuchsia-async took a dep on tokio in fxrev.dev/897738.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909851
Original-Revision: 9b7dee72e237b00f077536a21fbf85ed7e3ecaf6
GitOrigin-RevId: d9ecb3617356b8ea56705108cb95a237eae3c851
Change-Id: Iaba4d780200cdc401cd5d3760f27f0b2d93905fe