[roll] Roll fuchsia [tracing][rust] Elide strlen on CStr Conversion
CStr normally is represented by a slice, a 2 word combination of a
pointer and a length. When we call the `cstr!` macro for trace events,
we end up with 2 calls to strlen to generate a CStr. One for the the
name, one for the category.

The cstr! macro uses from_ptr, which has the option to be computed at
compile time, but is currently unstable. This swaps to using
CStr::from_bytes_until_nul which is const evaluable, dropping the need
for a runtime strlen. For an example trace point, e.g.

```
   trace_instant!(
        trace_category_starnix!(),
        trace_name_start_kernel!(),
        fuchsia_trace::Scope::Thread
    );
```
We now produce the code:
```
        movq    %rbx, (%rsp)
        leaq    .L__unnamed_979(%rip), %rdi
        leaq    .L__unnamed_980(%rip), %rdx
        leaq    .L__unnamed_879(%rip), %r9
        movl    $8, %esi
        movl    $12, %ecx
        xorl    %r8d, %r8d
        callq   _RNvCseWcUs29L097_13fuchsia_trace7instant@PLT
```

Original-Fixed: 124846
Original-Fixed: 124980
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892737
Original-Revision: 15ddb3728428c483ea5d3d8240c5231b50ddf1bd
GitOrigin-RevId: 2ec7fdc66ee92171031b507a139a3437a45cf59b
Change-Id: I4244fe8f963fd6b26c351e4c3089d830da9fa13c