[roll] Roll infra/recipes [go] Stop using /usr/bin/{gcc,ar,...}
These binaries are no longer found in Infra.

We can trivially disable cgo to workaround this as it is not necessary
to build the toolchain, however there are several compile tests that do
not respect the environment variable, so also pass `CC` and `AR`, and do
some more shenanigans to get `ar` on PATH because go also ignores the
`AR` environment variable sometimes too.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/909892
Original-Revision: 923dd244f696969d80612240808fc5b56ca94810
GitOrigin-RevId: ed5b8d0ff126352e6a4ea7e521f4d0ffac089c98
Change-Id: I8409325a08f354649c1801e26f4bd5e5a62d2f75