[roll] Roll fuchsia [fxtest][rewrite] Consistent handling of --count and --limit
We would miscount in the status display when the user specifies --count
or --limit. With this change we now display the number of test
executions that will be expected.

Additionally, the concept of "test was selected but is not run due to
user preference" is now exposed in the event data model for logs and
display purposes.

Original-Fixed: b/298207598
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909781
Original-Revision: bebcacc01ada3781eacb5c562257183809cb428f
GitOrigin-RevId: a1e7d82440700a56a3ca4dc187e83c96b2ca886f
Change-Id: I73402e3d4a2ac68f7076217e8f4162b1427f91cc