[roll] Roll fuchsia [blobfs] Remove fuchsia.blobfs.Blobfs
This API was never used and there are no plans to use it, thus there is
no reason to port it to Fxblob.  If we need an API for corrupt blob
handling in the future, we can create a new one.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910024
Original-Revision: df993dcb29e891e8bfaf1612aaa54210c333e5f8
GitOrigin-RevId: daccea59d844ea8db1e376769ebcb4946e70bdc0
Change-Id: I13c6ee804ed98d9b82a02a76f42d6ea7608836b0