[roll] Roll fuchsia [build][tests] Validator for test types
This is based on fxr/906930, but makes substantial-enough changes that
it has been created as a new CL.

The 'validate_test_type' tool takes a pair of files with the same
metadata as:
  - tests.json
  - test_components.json

and then using those files, can validate that all tests match some
criteria.

This initial version only supports the verification that all tests
are hermetic.

In order to do that without encountering its own errors, it contains
substantial bits of code required to understand how host tests are
represented, as well.

This is meant to be used with a GN template (in a following CL), which
can validate that all tests in a group are hermetic tests.

Original-Bug: b/297364940

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910320
Original-Revision: 8b1a285f1e4c4120d54815e23358c0fa872e4cd6
GitOrigin-RevId: d565f5aadb571054f8289422c9c4167ad2b33165
Change-Id: I7adbcdc8cab79319a485b8b077581026bc6349a2