[roll] Roll scudo [scudo][tests] Store the allocator instance in a global rather than posix_memalign it
The combined scudo allocator object is over 4MB in size which gets created via
the posix_memalign on every test run. If the tests are sanitized with asan,
then the asan allocator will need to mmap this large object every single time a
test is run. Depending on where this is mapped, we might not be able to find a
large enough contiguous space for scudo's primary allocator to reserve an arena.
Such a case is more likely to occur on 39-bit vma for RISCV where the arena size
is roughly a quarter of the whole address space and fragmentation can be a big issue.

This helps reduce fragmentation by instead placing the allocator instance in a
global storage rather than doing an anonymous mmap.

Differential Revision: https://reviews.llvm.org/D158767

GitOrigin-RevId: 1573b9405af3f64177406744d8270e65a415ed90
Original-Revision: b465fc76f458ca2c308ba12872e220d8eee04c1e
Change-Id: I137dd5b8eb49799a8b4b227b664fe2b4f6243bc9