[roll] Roll fuchsia [superproject] Roll scudo [scudo][tests] Store the allocator instance in a global rather than posix_memalign it
The combined scudo allocator object is over 4MB in size which gets created via
the posix_memalign on every test run. If the tests are sanitized with asan,
then the asan allocator will need to mmap this large object every single time a
test is run. Depending on where this is mapped, we might not be able to find a
large enough contiguous space for scudo's primary allocator to reserve an arena.
Such a case is more likely to occur on 39-bit vma for RISCV where the arena size
is roughly a quarter of the whole address space and fragmentation can be a big issue.

This helps reduce fragmentation by instead placing the allocator instance in a
global storage rather than doing an anonymous mmap.

Differential Revision: https://reviews.llvm.org/D158767

GitOrigin-RevId: 26be60c60c877a0e397031ae8a6953562957ca03
Original-Revision: b465fc76f458ca2c308ba12872e220d8eee04c1e
Roller-URL: https://ci.chromium.org/b/8771200863706045857
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910957
Original-Revision: 0fcb38d11e4e9a945627057657a76ace6e83c903
Change-Id: Ia2497bf3b7b7bdf5a88e6516c38926d552ee785d