[roll] Roll fuchsia [ffx][overnet] Refactor ssh to allow specifying the path
This adds the ability to specify the path to the ssh command
when building ssh command lines.

This ability is used in the host_pipe creation for overnet in order
to enable specifying a different command from `ssh` when testing
so the target side of the conversation can be easily mocked.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910953
Original-Revision: 9772d586ebe17e2f8fe175f0cc4a239ee5bf1c74
GitOrigin-RevId: 62c771582097243a21194307bfca4b902e948377
Change-Id: Ib04eb84d0941d00bf6ee0e0376974a6533728785