[roll] Roll fuchsia [audio] TRACE events for packet timing
Emit TRACE_INSTANT events when tracing is enabled for the
"audio" category, in certain per-packet circumstances that
are valid behavior but may not be the client's intention.
These checks are only active in tracing-enabled builds, so
their cost is only incurred when a developer effectively
opts-in on an engineering build intended purely for
diagnostics purposes.

Discrete events are emitted for each packet that is:
1) Timestamped, gap >= continuity threshold (pkt unchanged)
2) Timestamped, gap < continuity threshold (gap eliminated)
3) Timestamped, overlap < continuity threshold (no overlap)
4) NO_TIMESTAMP+DISCONTINUITY, received late (delay packet)
5) NO_TIMESTAMP+DISCONTINUITY, received on time (unchanged)

No change on non-tracing builds.

Test: caused these behaviors on TRACE build, observed events
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910833
Original-Revision: 3171cc4deacf88473c1511ec02b6730cb1d58700
GitOrigin-RevId: a26d134198d631115dbadb6576bc18314ba92006
Change-Id: Ie3e8e338231d4548b0d29c1df3905440bfeaf424