[roll] Roll fuchsia [ffx] Close Overnet connection when peer goes away
This fixes a bug in Overnet where the connection is kept
open even after the peer is lost. This causes issues
with ffx log trying to automatically reconnect to the device.
The daemon should wait for the device to come back up
before trying to reconnect to it.

Original-Bug: 132955
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910960
Original-Revision: e1b37465b6b8405ccae6df71774c75f7d803f8bf
GitOrigin-RevId: 0373b8295622764cd7ebbea7ac4ef7625db685bc
Change-Id: I455de9c2947755fc07143c59bf04d462546f95ee