[roll] Roll fuchsia [storage][perf] Run the host benchmarks on starnix
Run the host compatible storage benchmarks against the /data and /tmp
mount points in starnix.

starnix-data is a wrapper around fxfs so comparing the benchmark results
between the 2 filesystems gives a good indication of the cost of the
starnix layer.

memfs and starnix-tmp are both in-memory filesystems. starnix-tmp is
implemented directly in starnix while memfs is in its own component.
Comparing the OpenFile and StatPath benchmarks between the 2 filesystems
gives a good indication of the FIDL overhead of memfs and other fuchsia
filesystems.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903004
Original-Revision: 378db607dc374c4abd806284f7d0aef470941a41
GitOrigin-RevId: 47a19845035df518c9e8c2d28dbe88ef08f96610
Change-Id: I5528927ffe9c6822572939f299ae7ce49c786e4e