[roll] Roll fuchsia [libc] Add unmangled symbol in abi
This starts the transition period to use the unmangled symbol
name rather than the mangled which was accidentally not marked
`extern "C"`. For now, the old symbol is kept in the abi, but
we've excluded it from the ifs file so our link stub will not
have it. There should be an error if this symbol is used at
link time, though while we wait for the 6 week window it will
continue to be included in the libc.so used at runtime.

Original-Bug: 98888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669654
Original-Revision: 89c775615471c99fdb907d0557e72b25fbdbf950
GitOrigin-RevId: 1e798c66e5195721565f5f07deed9e50cafb2733
Change-Id: I55dca27f46bb850a80e434594bd4d1ce8c56f957