[roll] Roll fuchsia [diagnostics][fidl] Move host API to partner_internal
This API is used in `ffx inspect`, `ffx log` and other ffx plugins
that read Logs or Inspect from the device. The API is equivalent to
`fuchsia.diagnostics.ArchiveAccessor` (in the partner sdk), except
that it sends the data over a socket instead of a VMO as we can't send
VMOs to the host.

We wish to move it to partner_internal as part of having the ffx log
and ffx inspect tools in the SDK.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905520
Original-Revision: 63de69a4f9bd56ac91857b2e696faa3cbd6fb903
GitOrigin-RevId: 167f4f230f7c9e164ed74d7fed10c5b9d620b399
Change-Id: I7fb8e25dc9f3590d8af45507a036dfb6af7f2b9f