[roll] Roll fuchsia [sdk] Remove obsolete Rust merge program.
This removes a very dated Rust program that was intended to replace
//scripts/sdk/merge/merger.py, but was never properly completed or
used. The implementation:

- Has drifted considerably with many SDK metadata schema changes.
- Does not support newer features like fuchsia packages.
- Does not support merging multiple input directories into a single
  final export directory and/or archive.

It's better to remove this entirely, and focus future changes
on the Python script (which we may want to augment with better
tests though).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912032
Original-Revision: 25c4f1edf5786e81ee9f3273ef6a2ccac207ee74
GitOrigin-RevId: 9306db1211bffc43010e14426fd89b7676458d20
Change-Id: I06b5feb5e81b4b650fb52f7b830a6e4265b02d77