[roll] Roll fuchsia [ffx][rcs] Add ffx side of new rcs::open_capability
This adds support in the ffx libraries for using open_capability to
request a particular capability set instead of always getting exposed.
It leaves in shims on both sides so that if you *are* asking for
exposed, it uses connect_capability to avoid disrupting existing users,
and on the other end it leaves connect_capability functions that always
pass exposed to avoid having to change a lot of existing code.

This should mean everything that currently exists should continue to
work as-is, while small uses of open can start to be used (like the new
toolbox capability).

Original-Bug: b/292290062
Original-Bug: b/289268551
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903122
Original-Revision: b6ddcf391d2161b46dd540bc7e39052ddd49511c
GitOrigin-RevId: 2e88c3c700b832620b464e51da810e1b7df3d244
Change-Id: I9af83ce8b8d80fceb4fd5028e996bbb4cd7c1a2b