[roll] Roll fuchsia [fxtest][rewrite] Make error more clear
If someone adds a test to their args.gn but does not re-run the build,
the test will be added to tests.json but potentially not test-list.json.

This error message describes how to fix the problem if it happens.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912020
Original-Revision: 665524efd0f3c125f692e5f3f25ab9c5c4da614f
GitOrigin-RevId: 21fadf63790831c899444a09b9ba7b74db51a3a9
Change-Id: I7de18c649364c037a6130043278f63a50632055a