[roll] Roll fuchsia [test][vm] Fix race in Pager.EvictionHintsVmar
The EvictionHintsVmar test had previously been modified to account for
a race in a eviction potentially happening on a range marked DONT_NEED.
However, this handling was incompleted and only accounted for 2 out of
the 3 possible scenarios.

This change accounts for all 3 potential outcomes, and refactors the
handling into a single re-usable lambda.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911552
Original-Revision: bd5271795f143ab3b7c0816f8c83747a9877baf7
GitOrigin-RevId: 1a2233167b8b7a50eb8a47a3b895f2f624065201
Change-Id: Id47ff5d9aec2c7762961da0d955a927167d75b03