[roll] Roll fuchsia [virtualization] Add zircon guest smoke test package
Add a new test package, zircon-guest-smoke-tests, that simply boots a
zircon guest. This is intended to be the smallest zircon guest
integration test possible.

This test has a total size of 118MiB per fx blobstats. The biggest
contributors are the zircon guest images at 35MiB for the kernel and
53MiB for the filesystem.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910132
Original-Revision: 5648bb2e804678785ab878352572345250047b6b
GitOrigin-RevId: c24d2157a4e55c3dff16245885bf62941b059ddb
Change-Id: I5df98eb622c3a055b7f64930c985b9d377feb420