[roll] Roll fuchsia [trace] Add rust bindings for TraceObserver and ProlongedContext
These two APIs in tandem allow for programs to initialize a data source
when tracing starts, then collect data and write it to the trace buffer
when tracing ends. TraceObserver is used to know when tracing starts and
stops, while ProlongedContext is necessary to prevent the trace session
from terminating while the provider is in the process of collecting
the data to write.

BUG: b/271467732
BUG: b/291094354
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891187
Original-Revision: cd09f7a0b4a4f1019b988c7ed484ea40c9756252
GitOrigin-RevId: 4b0e4ba9b10057e16b279c88d8ef358dad3a7e99
Change-Id: I953247fb0ef9eb3e5f0914dcf3a81408a95cc321