[roll] Roll fuchsia [fdf][fidl] Create driver FIDL with fdf transport
Creates a new "Driver" FIDL protocol in the fuchsia.driver.framework
FIDL library that will be used as a replacement for the C based
lifecycle hooks in a DFv2 driver. The mechanism to connect to this is
through a single initialize C hook to exchange the channel that will be
used to communicate this new protocol.

Original-Bug: 130579
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884201
Original-Revision: 091da123493b3447ba8fbcee4b77383b8bf709d3
GitOrigin-RevId: f0deaac17d945aa908953b8541bde4405d67b359
Change-Id: Ie4d480a9ed0849fecc82005fae65167a7a6a8aea