[roll] Roll fuchsia [devshell][ffx] Use build-level target.default
This makes it so that we use the default target as a build level config
for ffx as the one set in the .device file adjacent to that.
Eventually we can remove the .device file as well, but for a while it's
good to do both for when people might be switching back and forth
between older and newer revisions a bit more.

As an extra precaution, we will check that if both configurations are
present, they are the same. If they differ, we suggest to the user that
they will need to re-run `fx set-device` to resolve the ambiguity. We
will return the ffx device in the meantime though.

This would most likely happen because they had switched to an older
version, run `fx set-device`, and then switched back to a newer
version without doing it again.

It also changes the warnings for running `ffx target default` to only
happen when there's likely to be a conflict, since get should generally
do the right thing now. It now only warns on setting, and then only if
you have a device pair file created.

Original-Bug: b/290948818
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821447
Original-Revision: 9fd4337a94e593add22a471122b2ae70c823590c
GitOrigin-RevId: f9d5c5addfdae3be1197712ac7ac8ccee6b294f7
Change-Id: Id50c6fdba10e5445214cbf344f1d9a847b94a23b