[roll] Roll fuchsia [netstack3] Garbage collection for neighbor table
When a neighbor entry transitions into a state that is considered
"unused" and therefore discardable (STALE or UNREACHABLE), if the size
of the neighbor table is above some maximum, schedule a timer to do a
garbage collection pass. The GC will attempt to remove enough neighbor
entries to bring the size of the table within the maximum. Static
neighbor entries and entries that are in use by the netstack (REACHABLE,
INCOMPLETE, DELAY, and PROBE entries) are never removed. UNREACHABLE
entries are removed first, then STALE entries.

Original-Bug: 35185
Original-Fixed: 132095
Test: netstack3-core-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910673
Original-Revision: 44db932317fad0a0db3d1095a6208f072df6d53d
GitOrigin-RevId: 44f077638d47b38529929710ced99a41dbe6552b
Change-Id: I8029eb81ea2f4b5b99e7942ac59a832928857987