[roll] Roll fuchsia [overnet][compatibility] abi-revision flag for overnet
This adds an optional flag `--abi-revision` to the
remote-control runner program which is used to start
the connection between a host tool and the target platform.

This revision is checked against the version info for the
platform and the result of the compatibilty check is returned.

It also merges in the returning of the $SSH_CONNECTION value to
the host tool.

If the new `--abi-revision` flag is not present, the legacy behavior
is done, in order to be backwards compatible.

Original-Bug: b/295899889
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912616
Original-Revision: e4669238ac3fbda35ec79c418fd069bf9eb8cd20
GitOrigin-RevId: 6feb2b36cd4046159ccbeb7208f8a9459d348511
Change-Id: I01ea6c60efcfc578e230354c318134d7bb2797b5