[roll] Roll fuchsia [sys] Serve each namespace entry as an Open/Directory capability
This is the minimal step to serving the namespace from bedrock
capabilities. We create an Open or Directory capability for each namespace
entry in the program, where the callback inside the Open just calls
the routing function. This revealed some tests that were relying on
a specific ordering of namespace entries which I corrected.

A later step would get rid of the routing functions altogether depending
on how many capability types are ported to senders.

Original-Bug: b/298106231

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909503
Original-Revision: b04416166b607792d285487c9da70ff6286e9936
GitOrigin-RevId: 789cbe9fd73dbaa6c60bc0734737220143697371
Change-Id: Ic26d7d58d565180d88081ab8aa66c4bf7b8008dc