[roll] Roll fuchsia [idk] Golden manifests over IDK.
Compute the "golden manifest" files for the IDK using the final IDK
itself, rather than the individual `sdk_collections` that make it up.
The `sdk_collections` are an implementation detail, rather than a
part of our contract with our partners - we should be validating
changes to the actual thing we ship to our partners, not intermediate
results.

The next step will be to update petals to use this new mechanism, and
then delete the existing `sdk_collection`-based manifest files.

I compared the new manifest file with the concatenation of core
and e2e_testing, and the only non-cosemetic change I found was that
arm64/aemu_internal and arm64/qemu_internal are listed in the old
manifests, but don't actually appear in the IDK, which I believe
is working as intended.

Interestingly, the tools for e2e_testing (fake-omaha-client and
trace2json) are also not built for arm64. Not sure if that's WAI or
not.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911952
Original-Revision: c13a930be4f336f5b4fe78ef8455c66ba9c30195
GitOrigin-RevId: 57c3ca477bf70aef78052278478b9ecfadf9947e
Change-Id: I990eafc9f0c364c68dcfedd86c57bb7943b536d2