[roll] Roll fuchsia [memory] Use memory maps to infer Starnix process memory
With this change, Starnix processes, identified by running in the same
job as a process named starnix-kernel.cm, will be attributed VMOs if,
and only if, this VMO appears in the restricted part of their address
space.

This brings the performance of Memory Monitor and ffx profile memory in
line with other systems when Starnix is used.

This change also removes the --include-starnix flag, as it no longer has
any meaning.

BUG=b/285191677

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903860
Original-Revision: f6fc826418691c398e0867ef9827bc2110d42140
GitOrigin-RevId: b8885d251f5a55aad16f035ac6b7c689b95492f7
Change-Id: I9d7f001115f3c9981d01138a8d557d3a7eee4539