[roll] Roll fuchsia [debroot] Include x11proto-scrnsaver-dev package
This is needed by SDL which is a dependency of QEMU. In newer releases,
this package was folded into x11proto-dev so this change only applies to
jessie and trusty.

In addition, we also switch the source URL to a faster mirror.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913372
Original-Revision: cf821ee2dbc2ed61dfc25a814dcc8d090f9fda03
GitOrigin-RevId: 1874b9fd61f82f2a48ecde85221f125fd8849015
Change-Id: I1f0962055098cbf388b52a2b316234eeb18f3dc7