[roll] Roll infra/recipes [build] Fix logic to use affected tests.
Coverage builders set affected_test_multiply_threshold=0 which was
causing -affected-tests not to be used in testsharder. Though it
specified affected tests in the modifiers.json instead, testsharder has
logic to not skip unaffected tests if -affected-tests is not provided.
This seems to have been left over from the migration to use
-affected-tests and testsharder should handle
affected_tests_multiply_threshold=0 properly, so we should be able to
just always use -affected-tests instead of specifying it in the
modifiers.

Original-Bug: 133238
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/913772
Original-Revision: 9ae17c91770934301305131d7b2ea15715e85fda
GitOrigin-RevId: 5c59f629766fa475f5ff6e15c566ce8515f014ae
Change-Id: Ic96de84bfd10ee10e12788cef66392f851a90749