[roll] Roll infra/recipes [shac] Avoid emitting duplicate comments
If a shac builder runs multiple times, it should not keep emitting the
same comments. To prevent that, we have the builder fetch the existing
robot comments on the patchset in question and skip adding any comments
that already exist.

We got this for free when using Tricium because Tricium only runs on new
patchset uploads.

Original-Bug: 82386
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/912736
Original-Revision: 16bfa8dd0d43a2ef907ece886112191ae7a9517e
GitOrigin-RevId: 5752ef5fa6a8f839cac226f85f0c10e67c15b3f0
Change-Id: I82a7afcb4b948068b2baedfc2eb7ee8b10af6859