[roll] Roll fuchsia [bt][snoop] Define an eager core shard
Currently, bt-snoop is started by bt-init when the device starts up.
This can be dynamically configured.
In some cases, only bt-snoop is required and not the rest of the
Bluetooth subsystem.

Define an eager core shard so that product integrators can use this in
cases where the rest of the BT subsystem is not included.

While here, update README with build instructions.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911254
Original-Revision: f6542a33c075664eee1fab79abf0e537f7e2b349
GitOrigin-RevId: 30052c72149eeaa919a1fb69b5942ce588675f35
Change-Id: I3048d2921651e3404ef6bc52abdb018ca4ab27f5