[roll] Roll fuchsia [netstack3] Don't keep Ethernet devices in IdMap
Remove the load from ethernet devices' scalar identifier from the IdMap
backing. Use a HashMap instead, keyed on the allocation given by the rc
types.

Move the debug identifier to a named field in EthernetDeviceState to
prevent misuse.

The Weak version of the identifier copies the ID out on downgrade so
it's accessible when debugging, but it's not used in any equality or
other load-bearing comparisons or collection searching.

Extract the state inside Loopback and Ethernet states into type aliases
to avoid the long and hard to read repetition.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913973
Original-Revision: b140122451e2933996ef9550789a14a4c8cf1811
GitOrigin-RevId: d3f65315f682438a9d1bc2a9a0df1542f6212c0a
Change-Id: If6eca788ba73ebccf71c7e83b16c448bdd125da9