[roll] Roll fuchsia [ld] Wire up MockLoaderService for use in tests
This introduces a new base class for the test fixtures used on
Zircon, which now all can send a loader service channel in the
bootstrap message backed by a mock loader service.  Tests can now
use `Needed({"liba.so", "libb.so"});` in between `Load("foo")`
and `Run()`.  This requires that the dynamic linker under test
request exactly those SONAMEs in that order, and answers them
with the VMOs of those names as found in /pkg/lib/test/lib.
(For later failure tests, there are also fixture methods to
expect requests that should return an error and no VMO.)

The Zircon LdStartupSpawnProcessTests fixture implicitly injects
an expectation for the file in the PT_INTERP string that the
system program launcher will request before transferring the
loader service channel to the new process.

This was tested manually by enabling LdStartupSpawnProcessTests
for Zircon locally and seeing that it succeeds in having the
system program launcher load the dynamic linker from PT_INTERP
via the mock loader service.  That set of tests remains disabled
because they now crash with stack overflow, but that's a separate
problem.  This testing demonstrates that things should work when
new dependency-loading tests are added and used in the other two
Zircon test fixtures.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912683
Original-Revision: 5f7776214cfbe020b90abbf10b0da4c248e09a0f
GitOrigin-RevId: 2b0b2d145d065377b4c735277eca48a7212442ee
Change-Id: I627687ba3a0f3b54705828b5eacdbe9a11229eab