[roll] Roll fuchsia [rtc][intel] Strip nvram support out of intel-rtc driver
The current intel-rtc driver has an improperly composed protocol of the
rtc and nvram protocols. The closed nature of the composed protocol is
making it difficult to do SDK-work on the rtc protocol at large. It
turns out that the nvram protocol isn't actually consumed by anything
other than the tests for the driver itself.

To remedy, I dropped the composed protocol and stripped out the nvram
interface of the intel-rtc driver. It is now just a provider of
ZX_PROTOCOL_RTC as per usual.

Fixes: 133214
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913705
Original-Revision: 0ef6deef06256115b339e2d595fcfa52da61dc3d
GitOrigin-RevId: 503fc511e8f9a5e45677a8b8a238e44226aa9f95
Change-Id: Id1d03a5bf31b9531bd998558105c6cc9d39cce0a