[roll] Roll fuchsia [netstack3] Yield to executor from device queue
Immediately signaling the waker from within the queue handler functions
did not effectively yield execution to other tasks. Instead of
scheduling the waker when hitting the batch size limits, return the
state of the work queue to the caller so it can yield once and call
again.

Introduce a core-wide concept of yielding in WorkQueueReport, which may
be used again for this pattern (like TCP do_send).

Original-Fixed: 132921
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913000
Original-Revision: ef9401c2980821db149eefd8dc3f6629119d2387
GitOrigin-RevId: 7a6b429841c7015cad9adc940129eb2d68a9e1a6
Change-Id: Ifde4f66a9509e10124113d73cc93ff0bb326197d