[roll] Roll fuchsia [starnix] Split process stop state
Previously, a thread group would enter a stop state as soon as
a stop signal was sent.  This CL changes that.

Thread groups are now considered in |stopping| state when a stop
signal is sent, and in |stopped| state when at least one of their
tasks has stopped.

The test test_stop_cond depended on the original stop behavior. It
was easier to replace it with another test than try to fix it. See
WaitpidFlagsTest.waitPidStopContinue.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914117
Original-Revision: 1f141a3bb58474ef1d785ef75206722c47c50ab3
GitOrigin-RevId: 8d7d4ed7baeede6b9827fbcfae289ef7e441e96b
Change-Id: Ia86545b731ce8bcf4abd64f7fc19f73093431f60