[roll] Roll fuchsia [netstack3] Test interfaces-admin using TestSetup
The interfaces-admin implicit_address_removal_on_interface_removal test
directly constructs Ctx in order to test some address-removal behavior
when a device is removed. This relies on Ctx being Default-constructible
without using a NetstackSeed, which will imminently not be the case
after a subsequent CL.

This change makes that test use the integration-test TestSetup
fixtures instead. As part of this change, we switch to using a
netdevice rather than loopback interface, as the latter are never
removed during normal netstack operation. In order to support using
netdevice interfaces for tests like this where we need to spy on
the `InterfaceEvent`s produced, we introduce a TestSetup mechanism for
copying all `InterfaceEvent`s sent to the worker so that they are
observable from tests as well.

Multiply: netstack3-bin-instrumented
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913152
Original-Revision: 2893049d16abba730b8442d861d4f75e6e635b55
GitOrigin-RevId: 5b7f9a1feca3de561e186ab345b27190734806b6
Change-Id: Idc76ef2fa424d6b2dbe4189b82ca4b86e3627207