[roll] Roll fuchsia [netstack3] Move routing table to bindings
Moves the source-of-truth for the routing table from core to bindings.
Now, instead of being able to add or remove routes directly, core will
dispatch events requesting that bindings add or remove routes from the
table.

This is favored over modifying routing table state directly in core,
because most routing changes originate in bindings, not in core, so it's
easier to manage locking and exclusive access to make route changes if
that state is held in bindings.

This also facilitates introducing bindings-specific notions of routing
table state, such as the reference-counted route sets specified by
fuchsia.net.routes.admin.

There's only one place in core that adds routes outside of unit tests:
NDP route discovery. Additionally, removing a link-layer device also
uninstalls all of the routes through that device.

This is a relatively large change in terms of linecount, but is hard to
reduce further due to the amount of coupling between how routing state
is held and how core and bindings manipulate and observe it through
events. Conceptually, this change does the following:

- Introduces a routing table in bindings.
- Adds event types that core can dispatch to tell bindings to update the
  table.
- Makes `AddableEntry` require the device to be specified in all routes.
  - The only place that was exercising device-selection-via-gateway was
    fuchsia.net.stack, so the stack_fidl_worker can be made to select
    the device first instead.
  - In the bindings route table, the AddableEntry will currently be held
    with a WeakDeviceId to avoid panics when core removes devices; see
    https://fxbug.dev/132990.
- Introduces an mpsc Sender to NonSyncCtx over which route change
  requests can be dispatched.
- Holds the corresponding mpsc Receiver in a separate route changes
  worker that executes all of the changes asynchronously and
  communicates back the success/failure of those changes over a oneshot
  channel.

Original-Bug: 126729
Original-Bug: 132990
Original-Fixed: 132470
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909875
Original-Revision: aaa920cf088e23b805058745351e7a4379a089b4
GitOrigin-RevId: de824b99ecd6b61ae24bef09624a572bb6d6ba31
Change-Id: Id4767c1c802e13bac7da28654856096000001e1e