[roll] Roll llvm-project/libc [libc] Implement stdio writing functions for the GPU port (#65809)
Summary:
This patch implements fwrite, putc, putchar, and fputc on the GPU. These
are very straightforward, the main difference for the GPU implementation
is that we are currently ignoring `errno`. This patch also introduces a
minimal smoke test for `putc` that is an exact copy of the `puts` test
except we print the string char by char. This also modifies the `fopen`
test to use `fwrite` to mirror its use of `fread` so that it is tested
as well.
GitOrigin-RevId: 8baf39bcfd11c6a12a6b77823e14f64ae121757d
Original-Revision: 80307938042d8edfc849663abe13c4deb983a8da
Change-Id: I315182673626c21201db8ab4a30ac5b916fefa6e