[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Implement stdio writing functions for the GPU port (#65809)
Summary:
This patch implements fwrite, putc, putchar, and fputc on the GPU. These
are very straightforward, the main difference for the GPU implementation
is that we are currently ignoring `errno`. This patch also introduces a
minimal smoke test for `putc` that is an exact copy of the `puts` test
except we print the string char by char. This also modifies the `fopen`
test to use `fwrite` to mirror its use of `fread` so that it is tested
as well.
GitOrigin-RevId: 1ed0d8c2e7bcd1527b875dbe19bef206ebcbda4f
Original-Revision: 80307938042d8edfc849663abe13c4deb983a8da
Roller-URL: https://ci.chromium.org/b/8770401383159187233
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914356
Original-Revision: 37a2df983bc657238a06e0de9bd67d3e9aa66118
Change-Id: If736fa2393053d6100fb35c78c59016ddb7eb9c6