[roll] Roll fuchsia [netstack3] Enable instrumented caller tracking
Replace StrongNames and all the previous machinery in the rc module with
a simple HashMap that tracks std::panic::Location instances and add
`[track_caller]` to the 3 methods that can create strong references.

This should still be the same no-op when not instrumented but now
there's actually the possibility of seeing all tracked clones when we
print.

Cleanup all the now unnecessary machinery to keep track of names and
update tests.

Original-Bug: 122388
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914497
Original-Revision: 835db2642c2a0758eaf1eed50a117aeba60ccb25
GitOrigin-RevId: 8f3d173f56c294a6d51033e9680bc9e7721bf759
Change-Id: I23ff7b74799709428d0a8e8a42e47642a1e7139b