[roll] Roll infra/recipes [rust_toolchain] Disable ThinLTO on beta builds
Use of LTO has led to miscompiles of our compiler in the past, so it's
not a great idea to use it for bootstrapping.

This won't take effect until the parent CL is merged and the builder
config is updated to set the "channel" property.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/907972
Original-Revision: e0e48afb8f8339e75080f69ec8417077948d34cb
GitOrigin-RevId: 28057ae98d80aec06488cca517b2b7b310df8836
Change-Id: Ib8962fa148f3d1fe96256ff448d97f37a28ed8c1