[roll] Roll fuchsia [kernel][vm] Don't call InitReclamation during PmmNode ctor
PmmNode's ctor is called super early as it's currently a global
variable.  Prior to this CL, the ctor called InitReclamation, which
assumes that some machinery is already setup and safe for use.  This
change zero initializes a few of PmmNode's fields and removes the call
to InitReclamation from the constructor.

Original-Fixed: 133162
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912258
Original-Revision: 230a19db20e8457a2ccdb2b0d80e42c9976dde6c
GitOrigin-RevId: cdab5af979c2ca5097cc67fbe7adb0f2e48e5df2
Change-Id: I22d3a47cfe4bfbffabadd6df0ee184e9206e5124