[roll] Roll fuchsia [netstack3] Prepare for Primary reference waiting
Simplify Primary a bit by removing the extra refs drop allowance.
Carrying that around was only necessary for `unwrap` and we can solve
that a bit more locally in `unwrap` without spreading the complexity
around.

Use ManuallyDrop inside Primary and Inner to allow unwrapping instead of
ptr_read. These are equivalent but ManuallyDrop is more purpose-built
for exactly this.

Add a test to check the double-panic path is protected and added a note
about unwind safety on Primary. The right thing to do here would
probably be to 'poison' the references and prevent further use, but we
don't need it.

In a follow-up, we'll allow unwrapping `data: T` from inner through a
notifier to close the loop on waiting for all references to drop.

Original-Bug: 122388
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914354
Original-Revision: 5a9871ff56bb6b1ebf9e4adbaf9ba69d33aea26b
GitOrigin-RevId: 88e72b3d60b7d2660feea32eabe1ef70cde70f1a
Change-Id: Iace84e003a1bf10e9e425598858bf01904a794d2