[roll] Roll fuchsia [tefmocheck] Fix log splitting to match test names.
go/bbid/8770228455612327809 is an example where a line matched the
testFinishedRegex but wasn't the line we were looking for so it later
failed to process because the test name we parsed from that line didn't
match the test names in the summary.json. We should check that the test
name we parse from the line matches the expected test name; otherwise,
skip that line.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914956
Original-Revision: 264804a74e053617f7d7e222352f267f1a236f21
GitOrigin-RevId: 1391d59a580caddb6fd5b9d9943a9ac0ff2f6e3c
Change-Id: I6f9e3c780abbd5dba16604b5df84944640b07bf4