[roll] Roll fuchsia [realm][fuzzing] Non-production fuzz targets
This CL replaces `testonly = true` with a `dep` on
"//build/validate:non_production_tag" for source sets that may be linked
into fuzzing targets. These targets are real code that likely is not
test-only, but fuzzable versions of them should not be used in
production.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907725
Original-Revision: 7811982832e0d9560eacc103f204454e7ce221d9
GitOrigin-RevId: 8c2691d96d5aa59e507918a42ba9a84a6f46e28e
Change-Id: Id6b7665c49a3ab7840de425c1083497a8c94d53e