[roll] Roll infra/recipes [icu] Adds the `ref` parameter to params
The old code assumed the ref `latest`. But in the
case of Chrome, this ref is not used at all.

So, instead of guessing, introduce a parameter.

An unspecified parameter resolves to a chromium ref as used in
Fuchsia.

Original-Bug: 300024441
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/915452
Original-Revision: c8448b1c46c4de9f3ed6602d0732cc2263f3b8d3
GitOrigin-RevId: 80f6335a1ec98ea2a649ce1f43db5aa211e11532
Change-Id: Id65f0110d1c4f265bda6601d7dbb15ed17fbbb45