[roll] Roll fuchsia [sampler] Simplify Inspect selectors in RealmProxy-driven test
We can exploit the fact that we have an archivist running in the test
realm that receives scoped events for the test case. That archivist
just sees "sampler" as the child, so we can get that directly and remove
the leaky abstraction that reveals internal test topology details.

Additionally, we now configure the sampler component name instead of
the realm builder instance name as that's the topology bit that is
visible and relevant to the test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915375
Original-Revision: 5273a57b65f4f280e3a27c7b1759803cec77ee78
GitOrigin-RevId: a2a71f10b8f6c16377f7dd6b489facbb3d0e6308
Change-Id: I3c946b7cb8a6163e20bcde8b9c3e7f198bf34c1a