[roll] Roll fuchsia [netstack3] Delete is_device_installed
The only place this was used was for eligibility for device selection
from routes. Disable IP on that device has the exact same effect, so
doing that on removal.

In the world where core stops trying to remove device routes by itself
and delegates this cleanup to bindings, this is likely irrelevant but
serves as a statement that as long as bindings itself doesn't attempt to
re-enable IPv[4|6] on that interface it shouldn't accept new addresses
or packets.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913901
Original-Revision: 6e31804f5303e5f320f47bafcbbcb63d481caaae
GitOrigin-RevId: 4cf92301ce31ec1031da12ad722d2e8772f5864d
Change-Id: I0ca867c565352b645a6818bfefda5f23ec6fa0a8