[roll] Roll fuchsia [amlogic-dipslay][hdmi] Do not alter the display mode.
Previously the amlogic-display driver altered the display mode
to support 4K resolutions. The driver had to keep an unaltered
display mode for comparison, but use the altered display mode to
drive the display hardware. This workaround was never used in
amlogic-provided display driver software.

This implementation doesn't work for 4K displays (see bug
fxbug.dev/124984) and it also reduces the code readability.
Now that we are not targeting 4K displays, we should be able
to remove this hack so that the display mode will be always
passed through.

Test: display-tool squares 1 on VIM3 using the following displays:
      [1] MIMO display (1280x800).
      [2] Google JN32A (3840x2160) with resolution fallback to 2560x1440.
      [3] Acer B326 (3840x2160), with resolution fallback to 1920x1080.
      [4] Dell P2720DC (2560x1440).

Original-Bug: 124984
Original-Bug: 132602
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912261
Original-Revision: 3f3dda3e1b5dc229b50e45334406710683a6e92d
GitOrigin-RevId: 03c9c4c401df24327c1b7e1cd05070c3b5b139a6
Change-Id: I664d26b078eef6b7e21269f42cd2170186247465