[roll] Roll fuchsia [ffx][emu] Improved error message when emulator not supported
This updates the error message to be more clear when the product bundle
does not contain virtual device specifications. These virtual devices
are used to configure the emulator. The are not present in product
bundles that do not support running in the emulator, such as product
bundles that use a physical board type that does contain emulation
support.

Original-Bug: b/294271932
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915992
Original-Revision: c73180758600cfafb89384767a64c7c2830c54a7
GitOrigin-RevId: 39a1b0b1027170b0ef7491547357edc292e00d43
Change-Id: I8ff5200bbf552035aa7cde2e70728211337fcf36