[roll] Roll fuchsia [debugger] Improve process to component mapping
Component instances and processes are not necessarily a 1:1 mapping,
although this is common for components which use the ELF runner. Any
component instance may spawn many processes or, many component instances
may map to a single process (i.e. driver-hosts). It is also possible for
a component to purely have routing information and/or host children and
be related to no process at all.

debug_agent's understanding of components and processes assumed that a
job would have a single component associated with each child process.
This allowed users to attach to components by moniker, URL, or any
subset of the two. However; when a particular job houses many
components, this 1:1 assumption caused filters to not match against all
components in a job or associated with a process, making certain systems
(drivers, in particular) challenging to attach to.

debug_agent now has a more correct mapping of N components for a
particular job, so new filters will be able to search across all
components for the given job.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915612
Original-Revision: d7b0ff48a17d175379d2ee94d1a57e8de271ea28
GitOrigin-RevId: fb37b04ab3bbf4343058963044e93e0e5b2f6cbb
Change-Id: I88191fda527e2f53307bdbccfb395151549e137a