[roll] Roll fuchsia [assembly][time] Allow timekeeper configuration
Some tests require that the UTC clock is available and running,
while there is no network available.

Since the Fuchsia UTC clock can not function in that case,
we configure timekeeper with a feature flag that starts the
UTC clock regardless.

Most of this change is plumbing the configuration bits to the
right place.

Original-Bug: b/295537827
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915695
Original-Revision: 71716fa03018d32852915c0014a37ea6132c7dce
GitOrigin-RevId: bf184f69313b6846acf7f9408c9ad81b4ed841e7
Change-Id: I79d5bbac190bafe67df343738d3837bb018b702e