[roll] Roll fuchsia [overnet] Improve ascendd socket error reporting
This CL improves the error reporting logic when failing to bind the
ascendd socket. Before, it would assume that bind() only fails when the
socket is already in use, and tried to connect to that existing socket.
Now, it only does that on EADDRINUSE, and reports the original bind()
error otherwise.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916268
Original-Revision: caf941d793969df6d25b6a627dbd6366635c750b
GitOrigin-RevId: ff6347e57fe38cc5547617f400c712cb1a6f9c55
Change-Id: Ib88a8b3dc9f4534b18997b67ae67362c9202ff2b