[roll] Roll fuchsia [pw_rpc] Client/server component integration prototype
Add a prototype to //examples/components that demonstrates integration
between the Component Framework and an offloaded program that uses
pigweed.

The offloaded program is represented in Fuchsia as a component. This
component uses a 'pigweed runner' that bridges between Fuchsia and the
offloaded program (in this example, assumed to run on a Linux host).

The pigweed side change landed in
https://pigweed-review.git.corp.google.com/c/pigweed/pigweed/+/167158.

Original-Bug: b/299358404

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878075
Original-Revision: 2f32ea05353e27be9e6f3d102565e2bb2c957cc5
GitOrigin-RevId: d8a1bd1f4103594df25a34a468e68314a7b0dc47
Change-Id: Ic8130c5a097611fcf951ff7de7fefe81cef18907