[roll] Roll fuchsia [netstack3] Lift DeviceId Display impl
This type is too complicated to call its implementation Display, this is
mostly historical artifact from when it was a scalar. This guarantees
that we're using Debug impl consistently and provides guardrails for
diverging impls that are confusing to pick between.

While I'm here, remove Display impl for internal value that can be well
served by comments and use a Debug impl instead.

Original-Bug: 133595

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917174
Original-Revision: e0503a7cdd777d3757e433198ecdbe2a8409ccfb
GitOrigin-RevId: bbd44033607f8c95cabab41e9a6588ac1fea528f
Change-Id: I154865cbb79b20a58d319f299d3140260c92b779