[roll] Roll fuchsia [fuchsia-component] Support opening from a directory endpoint
Add support for opening protocols, files, directories, relative
to a directory endpoint, using fdio. This is useful when one only has
access to a &ClientEnd<DirectoryProxy>, hence cannot convert it to a
directory proxy.

fuchsia-fs is used to open these relative to a DirectoryProxy.
fxrev.dev/915818 was an attempt to add these to fuchsia-fs, but the
owners would prefer to avoid fdio in that library.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917139
Original-Revision: f10939fe079d9e98fcdb2610177320e56b823327
GitOrigin-RevId: d2d65d11e4b23325eb026b74d40f1aebeba0c169
Change-Id: I6cc9a2fb097cda354c11c35455d820a9c3ef24b3