[roll] Roll fuchsia [bt][avrcp] Remove duplicate type definition
There are 3 definitions of PlayerApplicationSettingAttributeId. The
first is a FIDL type definition in the `fuchsia.bluetooth.avrcp` FIDL
library. The remaining 2 are defined locally in the bt-avrcp component
and both are identical. It is valid to convert from the FIDL type to the
crate local type.

Remove the erroneous duplicate definition.

Original-Bug: 133584
Test: fx build && fx test bt-avrcp-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917021
Original-Revision: c5c72f3d79e9c5902fe5825dcdbdf2927284060e
GitOrigin-RevId: 87f398795d92612e45e1ca34f14c356c63ecbc25
Change-Id: Iaa74a3112d5b3837144eef3bbed030e62201398c